refactor tests a bit..
it looks like not having a mock rest interface for neo4j and the way
that dispatch is async is causing some issues with unit testing, causing
certain tests to outrun other tests, causing tests to intermittently
fail.  I added a Thread.sleep to test this and see if i can get things
passing on travis.  If this is infact the issue, might want to look and
see if there is anything like concurrent.Eventually that can extend
FlatSpec.