Rollback use of macro; split version into its own.
Using macro to guard against asynchronous code in block can work, but
only when the macro is passed the block directly, not through a
variable.  So it would involve providing withTx as a macro itself, which
then rewrites the block into its current form after checking.  But that
introduces circular dependency.  Letting the user add the check himself
is a simpler approach.